+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 ## 0.43
 - Speed up the creation of a dx:workflow. With one API call create
 all the workflow stages, and add set the checksum property.
+- Speeding up applet construction
 
 ## 0.42
 - WDL objects: experimental